refactor(topology): extract interface for ClusterTopologyManager
